+++ .gitignore
@@ -1,5 +1,6 @@
 .env
 __pycache__/
+*/__pycache__
 test.db
 .pytest_cache
 .idea
