--- alembic/versions/6364dbabd34e_sex.py
@@ -1,42 +1,0 @@
-"""sex
-
-Revision ID: 6364dbabd34e
-Revises: 
-Create Date: 2023-12-14 22:36:42.740681
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = '6364dbabd34e'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('ratings',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('rate', sa.Integer(), nullable=True),
-    sa.Column('user_id', sa.Integer(), nullable=True),
-    sa.Column('image_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['image_id'], ['images.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.add_column('users', sa.Column('sex', sa.String(length=7), nullable=False))
-    op.add_column('users', sa.Column('avatar', sa.String(length=255), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'avatar')
-    op.drop_column('users', 'sex')
-    op.drop_table('ratings')
-    # ### end Alembic commands ###
