--- alembic/versions/b1db0490444c_qr.py
@@ -1,30 +1,0 @@
-"""qr
-
-Revision ID: b1db0490444c
-Revises: 6364dbabd34e
-Create Date: 2023-12-15 12:16:48.776012
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'b1db0490444c'
-down_revision: Union[str, None] = '6364dbabd34e'
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.add_column('images', sa.Column('qr_url', sa.String(length=255), nullable=True))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('images', 'qr_url')
-    # ### end Alembic commands ###
