+++ alembic/versions/bad2fd3eae9b_init.py
@@ -1,0 +1,36 @@
+"""Init
+
+Revision ID: bad2fd3eae9b
+Revises: 
+Create Date: 2023-12-15 20:32:40.621002
+
+"""
+from typing import Sequence, Union
+
+from alembic import op
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+# revision identifiers, used by Alembic.
+revision: str = 'bad2fd3eae9b'
+down_revision: Union[str, None] = None
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.drop_table('token_black_list')
+    op.drop_column('images', 'image_name')
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.add_column('images', sa.Column('image_name', sa.VARCHAR(length=150), autoincrement=False, nullable=True))
+    op.create_table('token_black_list',
+    sa.Column('access_token', sa.VARCHAR(length=255), autoincrement=False, nullable=False),
+    sa.Column('created_at', postgresql.TIMESTAMP(), autoincrement=False, nullable=True),
+    sa.PrimaryKeyConstraint('access_token', name='token_black_list_pkey')
+    )
+    # ### end Alembic commands ###
