--- alembic/versions/bd5b1dfa5936_init.py
@@ -1,40 +1,0 @@
-"""Init
-
-Revision ID: bd5b1dfa5936
-Revises: 
-Create Date: 2023-12-15 15:28:26.486271
-
-"""
-from typing import Sequence, Union
-
-from alembic import op
-import sqlalchemy as sa
-
-
-# revision identifiers, used by Alembic.
-revision: str = 'bd5b1dfa5936'
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table('ratings',
-    sa.Column('id', sa.Integer(), nullable=False),
-    sa.Column('rate', sa.Integer(), nullable=True),
-    sa.Column('user_id', sa.Integer(), nullable=True),
-    sa.Column('image_id', sa.Integer(), nullable=True),
-    sa.ForeignKeyConstraint(['image_id'], ['images.id'], ondelete='CASCADE'),
-    sa.ForeignKeyConstraint(['user_id'], ['users.id'], ondelete='CASCADE'),
-    sa.PrimaryKeyConstraint('id')
-    )
-    op.add_column('users', sa.Column('sex', sa.String(length=7), nullable=False))
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_column('users', 'sex')
-    op.drop_table('ratings')
-    # ### end Alembic commands ###
